chore: update links to indexeddb-examples
As part of the repository triage we have been doing, we have been updating links to the repositories and live examples in /content. This can be easy for you all to miss, so I will open some pull requests to update these links as appropriate. This is the first of those.

fix #8857