fr: improve `/Learn_web_development/Core/Scripting/Object_basics` (#24268)
* Update index.md

improve french translation of 'what is "this"?' section of Javascript object basics

* Fast forward link change

* Revise JavaScript object basics content in French

Updated the content to improve clarity and correct minor errors in the French translation. Adjusted syntax examples and explanations for better understanding of JavaScript object basics.

* Enhance object basics with constructors explanation

Added a section on constructors and object creation in JavaScript, explaining their importance and providing examples.

* Update content on object basics in JavaScript

Removed mention of upcoming article on object-oriented programming.

* Fix formatting and update notification example

* Update object syntax example in index.md

* Update object name in example

---------

Co-authored-by: Tristan D. <tristantheb@users.noreply.github.com>