Fix important misconception about Date parsing behaviour wrt "invalid" dates
The sentence about the `Date.parse` method (hence also about the Date constructor) was saying that invalid dates (as in, dates that do not exist like February 31th) were detected and returned `NaN`. This is simply not true, and very easy to check using the very example that was used here (2015-02-31). This commit fixes that introductory sentence and adds a warning block about that frequent mistake.